initial remodel of app structure
