+++ frontend/spotted-apple/app/features/home-links-feature/schema.tsx
@@ -1,0 +1,6 @@
+export type CustomLink = {
+    key: string;
+    name: string;
+    href: string;
+    description: string;
+};
