+++ frontend/spotted-apple/app/lib/fonts/index.tsx
@@ -1,0 +1,3 @@
+import { Inter } from 'next/font/google';
+
+export const inter = Inter({ subsets: ['latin'] });
