+++ frontend/spotted-apple/app/lib/globals.css
@@ -1,0 +1,91 @@
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
+
+:root {
+  --foreground-rgb: 0, 0, 0;
+  --background-start-rgb: 214, 219, 220;
+  --background-end-rgb: 255, 255, 255;
+}
+
+@media (prefers-color-scheme: dark) {
+  :root {
+    --foreground-rgb: 255, 255, 255;
+    --background-start-rgb: 0, 0, 0;
+    --background-end-rgb: 0, 0, 0;
+  }
+}
+
+body {
+  color: rgb(var(--foreground-rgb));
+  background: linear-gradient(
+      to bottom,
+      transparent,
+      rgb(var(--background-end-rgb))
+    )
+    rgb(var(--background-start-rgb));
+}
+
+.createAccount{
+  padding: 1rem;
+  background-color: rgb(236, 234, 234);
+  align-items: center;
+  width: 50%;
+  border-radius: 7px;
+
+  input {
+    color: rgb(0, 0, 0);
+    text-align: left;
+    border-radius: 7px;
+    margin: .5rem;
+    padding: .75rem;
+    height: 2.75rem;
+    width: 100%;
+  }
+  
+  p {
+    color: rgba(0, 0, 0, 0.59);
+    padding: .15rem;
+    margin: .5rem;
+  }
+
+  .errorField {
+    font-size: 0.75rem;
+    color: rgba(250, 7, 7, 0.75);
+  }
+
+}
+
+.setPassword{
+  padding: 1rem;
+  background-color: rgb(236, 234, 234);
+  align-items: center;
+  width: 50%;
+  border-radius: 7px;
+
+  input {
+    color: rgb(0, 0, 0);
+    text-align: left;
+    border-radius: 7px;
+    margin: .5rem;
+    padding: .75rem;
+    height: 2.75rem;
+    width: 100%;
+  }
+}
+
+button {
+  background-color: rgb(37, 61, 220);
+  text-align: center;
+  border-radius: 7px;
+  margin: .5rem;
+  padding: .75rem;
+  height: 2.75rem;
+  width: 100%;
+}
+
+@layer utilities {
+  .text-balance {
+    text-wrap: balance;
+  }
+}
