--- frontend/spotted-apple/app/ui/fonts.ts
@@ -1,3 +1,0 @@
-import { Inter } from 'next/font/google';
-
-export const inter = Inter({ subsets: ['latin'] });
